biomecode-api/requirements.txt
fastapi
uvicorn[standard]
earthengine-api
python-dotenv
geojson
